Refactor actuator parser utility
Elevate ActParser into a core utility as it can be used in other places.
